Fix task saving and some cleanup
