+++ src/app/tasks/[slug]/edit/page.tsx
@@ -2,8 +2,8 @@ import { notFound, redirect } from "next/navigation";
 import { TaskEditor } from "client/components/task_editor/task_editor";
 import { getSession } from "server/sessions";
 import { getEditorTask } from "server/logic/tasks/get_editor_task";
-import { uuidToHuradoID } from "common/utils/uuid";
-
+import { huradoIDToUUID, uuidToHuradoID } from "common/utils/uuid";
+import { db } from "db";
 
 type TaskEditPageProps = {
   params: {
@@ -16,13 +16,23 @@ export default async function TaskEditPage(props: TaskEditPageProps) {
   if (session == null || session.user.role != "admin") {
     return { errorCode: 403 };
   }
+  const uuid = huradoIDToUUID(props.params.slug);
+  if (uuid == null) {
+    const task = await db
+      .selectFrom("tasks")
+      .select("id")
+      .where("slug", "=", props.params.slug)
+      .executeTakeFirst();
 
-  const hid = uuidToHuradoID(props.params.slug);
-  if (hid !== props.params.slug) {
+    if (task == null) {
+      return notFound();
+    }
+
+    const hid = uuidToHuradoID(task.id);
     return redirect(`/tasks/${hid}/edit`);
   }
 
-  const task = await getEditorTask(props.params.slug);
+  const task = await getEditorTask(uuid);
 
   if (task == null) {
     return notFound();
